clean up imgui event handling
