+++ src/utils/mod.rs
@@ -11,7 +11,6 @@ use smithay::{
 
 use crate::{framework::surface_data::SurfaceData, window::WindowSurface};
 
-pub mod imgui_input;
 mod iterators;
 
 pub use iterators::{VisibleWorkspaceIter, VisibleWorkspaceIterMut};
